KOGITO-1798: Disable example failing tests with persistence until defects are fixed
