+++ process-quarkus-example/pom.xml
@@ -129,6 +129,19 @@
           <version>${project.version}</version>
         </dependency>
       </dependencies>
+      <build>
+        <plugins>
+          <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-surefire-plugin</artifactId>
+            <configuration>
+              <systemPropertyVariables>
+                <tests.category>persistence</tests.category>
+              </systemPropertyVariables>
+            </configuration>
+	      </plugin>
+        </plugins>
+      </build>
     </profile>
     <profile>
       <id>events</id>
