+++ process-springboot-example/pom.xml
@@ -130,6 +130,19 @@
           <version>2.1.7.Final</version>
         </dependency>
       </dependencies>
+      <build>
+        <plugins>
+          <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-surefire-plugin</artifactId>
+            <configuration>
+              <systemPropertyVariables>
+                <tests.category>persistence</tests.category>
+              </systemPropertyVariables>
+            </configuration>
+	      </plugin>
+        </plugins>
+      </build>
     </profile>
   </profiles>
